fix(api, ui): 0% of sector emissions on report sector table (#1806)
Also fixes that some of the rows in the table weren't full width by using React.Fragment instead of <> or , so the key attribute can be supplied to it but the rows are direct children of the table body.
Also fixes N/A data sources being linked and not being translated.