fix(deps): install prettier v3.6.2 to match @react-email/render dependency
Fixes error in
./node_modules/@react-email/components/node_modules/@react-email/render/dist/node:
Package prettier can't be external
The request prettier/standalone matches serverExternalPackages (or the
default list).
The package resolves to a different version when requested from the
project directory (3.2.5) compared to the package requested from the
importing module (3.6.2).
Make sure to install the same version of the package in both locations.